bpfman: cli: Refactor command handling
This moves all clap code into args.rs to allow it to be used via the
include! macro in an xtask to generate manpages.

Updates: #703
